Moved out hooks from component to service.
